// src/app/shared/models/user.model.ts
export class User {
  id: number;
  firstName: string;
  lastName: string;
  gender: string;
  role: string;
  userName: string;
  password: string;
  departmentId: string;
  departmentName: string;

  constructor() {
    this.id = null;
    this.firstName = null;
    this.lastName = null;
    this.gender = null;
    this.role = null;
    this.userName = null;
    this.password = null;
    this.departmentId = null;
    this.departmentName = null;
  }
}